+++ apps/web/preload/script/api/index.ts
@@ -1,9 +1,13 @@
 import { processDom } from './dom';
 import {
-    getChildrenCount, getElementAtLoc, getElementByDomId, getOffsetParent,
+    getChildrenCount,
+    getElementAtLoc,
+    getElementByDomId,
+    getOffsetParent,
     getParentElement,
     updateElementInstance
 } from './elements';
+import { groupElements, ungroupElements } from './elements/dom/group';
 import {
     getActionElement,
     getActionLocation,
@@ -11,13 +15,15 @@ import {
     getFirstOnlookElement,
     setElementType,
 } from './elements/dom/helpers';
-import { getInsertLocation } from './elements/dom/insert';
+import { insertImage, removeImage } from './elements/dom/image';
+import { getInsertLocation, insertElement, removeElement } from './elements/dom/insert';
 import { getRemoveAction } from './elements/dom/remove';
-import { getElementIndex } from './elements/move';
+import { getElementIndex, moveElement } from './elements/move';
 import { drag, endAllDrag, endDrag, startDrag } from './elements/move/drag';
 import { getComputedStyleByDomId } from './elements/style';
 import { editText, startEditingText, stopEditingText } from './elements/text';
 import { setFrameId } from './state';
+import { updateStyle } from './style';
 import { getTheme, setTheme } from './theme';
 
 export const preloadMethods = {
@@ -58,6 +64,16 @@ export const preloadMethods = {
     startEditingText,
     editText,
     stopEditingText,
+
+    // Edit elements
+    updateStyle,
+    insertElement,
+    removeElement,
+    moveElement,
+    groupElements,
+    ungroupElements,
+    insertImage,
+    removeImage,
 }
 
 export type PreloadMethods = typeof preloadMethods;
